requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# Machine Learning & Deep Learning
tensorflow==2.15.0
xgboost==2.0.3
lightgbm==4.1.0

# Data Collection & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==4.9.3

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# sqlite3 is built into Python, no need to install

# API & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Jupyter for development
jupyter==1.0.0
ipykernel==6.27.1 